chore: extract handler from get
